Add CLI tool with pretty printing
This removes some of the work with pretty printing things in the various
modules. Unfortunately, it's also a pain because the import is weird.
I'm thinking about solutions to this. Probably, everything will end up
in a single directory at some point and I can just forget all the path
troubles.