Add top level cz_search file
This will be the entry point for the CLI, if I get around to making
that. At the moment, it imports the modules and queries Forvo, and
that's pretty much it.

Now that all of this is working somewhat, I can try to really turn it
into something useful and neat.